separated process execution code into `ProcessExecutor`
