+++ cli/CMakeLists.txt
@@ -39,9 +39,7 @@ endif()
 if(tinyxml2_FOUND AND NOT USE_BUNDLED_TINYXML2)
     target_link_libraries(cppcheck ${tinyxml2_LIBRARIES})
 endif()
-if (USE_THREADS)
-    target_link_libraries(cppcheck ${CMAKE_THREAD_LIBS_INIT})
-endif()
+target_link_libraries(cppcheck ${CMAKE_THREAD_LIBS_INIT})
 
 add_dependencies(cppcheck copy_cfg)
 add_dependencies(cppcheck copy_addons)
