+++ cmake/printInfo.cmake
@@ -52,9 +52,7 @@ if (HAVE_RULES)
 endif()
 message( STATUS )
 message( STATUS "USE_THREADS =           ${USE_THREADS}" )
-if (USE_THREADS)
-    message( STATUS "CMAKE_THREAD_LIBS_INIT = ${CMAKE_THREAD_LIBS_INIT}" )
-endif()
+message( STATUS "CMAKE_THREAD_LIBS_INIT = ${CMAKE_THREAD_LIBS_INIT}" )
 if (NOT USE_MATCHCOMPILER_OPT MATCHES "Off")
     message( STATUS )
     message( STATUS "PYTHON_VERSION_STRING = ${PYTHON_VERSION_STRING}" )
