+++ test/CMakeLists.txt
@@ -25,9 +25,7 @@ if (BUILD_TESTS)
     if(tinyxml2_FOUND AND NOT USE_BUNDLED_TINYXML2)
         target_link_libraries(testrunner ${tinyxml2_LIBRARIES})
     endif()
-    if (USE_THREADS)
-        target_link_libraries(testrunner ${CMAKE_THREAD_LIBS_INIT})
-    endif()
+    target_link_libraries(testrunner ${CMAKE_THREAD_LIBS_INIT})
 
     if (NOT CMAKE_DISABLE_PRECOMPILE_HEADERS)
         target_precompile_headers(testrunner PRIVATE precompiled.h)
